glossary/requirements.txt
bs4
ruamel.yaml
